refactor: exctract utils
