+++ modules/market/src/react-app-env.d.ts
@@ -12,3 +12,4 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 /// <reference types="react-scripts" />
+declare module 'resolve-pathname';
